% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sig.R
\name{find_sig}
\alias{find_sig}
\title{find_sig: Find residual variance of your outcome given a vector of coefficients}
\usage{
find_sig(n, cov_mat, beta, target_var_y = 1, iter = 10000)
}
\arguments{
\item{n}{Number of Obseravtion to be simulated at each iteration.}

\item{cov_mat}{Covariance matrix defining how X1 and X2 are related along with their scale.}

\item{beta}{Vector of coefficients mapping terms X1, X2, X1^2, X2^2, and X1*X2 to y.}

\item{target_var_y}{The expected variance of y.}

\item{iter}{The number of iterations to run to estimate the residual variance.}
}
\value{
Value of residual variance given proposed model
}
\description{
While the principles of variance algebra hold for linear relationships, they are complicated by non-linear terms and interactions. 
In fact, closed form solutions for the variance of product distributions is an active research area. 
This function capitalizes on the power of simulation to estimate the residual variance of a response surface given
the expected value of the outcome variance.
}
\examples{

# Defining Correlation Matrix describing how x1 and x2 are related
# Covarince and variance of x1^2, x2^2, and x1*x2 follow from this matrix
cov_mat<-matrix(c(1, 0,
                  0, 1), byrow = TRUE, 2, 2)
# Defining betas x1, x2, x1^2, x2^2, and x1*x2
beta<-c(0, 0, -.075, -.075, .15)

# Simulating 10,000 draws of size 1000 assuming the correlation structure and regression weights defined above.
sig_hat <- find_sig(n = 1000, cov_mat = cov_mat, beta = beta, target_var_y = 1)                 
sig_hat
}
