% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_response_surface_x.R
\name{gen_response_surf_x}
\alias{gen_response_surf_x}
\title{gen_response_surface_x: Generate IVs for Response Surface}
\usage{
gen_response_surf_x(n, cov_mat, x_names = NULL)
}
\arguments{
\item{n}{Number of observations to simulate.}

\item{cov_mat}{Covariance matrix defining how X1 and X2 are related along with their scale.}

\item{x_names}{Character vector of length two containing the names of X variables (optional)}
}
\value{
Data frame of IVs with requested column names.
}
\description{
A helper function for simulating the independent variables for a response surface. Automatically calculates the interaction and quadratic terms. 
For simplicities sake, assumes that variables have mean 0
}
