% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_response_surface_y.R
\name{gen_response_surf_y}
\alias{gen_response_surf_y}
\title{gen_response_surf_y: Generate DV under response surface model}
\usage{
gen_response_surf_y(x_data, beta, sigma = NULL, y_name = NULL)
}
\arguments{
\item{x_data}{IV data typically generated by gen_response_surf_x}

\item{beta}{vector of regression weights in the order of x_data}

\item{sigma}{error variance}

\item{y_name}{name of y variable}
}
\value{
data frame with x and y variables.
}
\description{
gen_response_surf_y: Generate DV under response surface model
}
