% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sig.R
\name{find_sig}
\alias{find_sig}
\title{find_sig: Find residual variance of your outcome given a vector of coefficients}
\usage{
find_sig(n, cor_mat, beta, target_var_y = 1, iter = 10000)
}
\arguments{
\item{n}{Number of Obseravtion}

\item{cor_mat}{Correlation matrix of IVs}

\item{beta}{vector of coefficients mapping terms X1, X2, X1^2, X2^2, and X1*X2 to y}

\item{target_var_y}{the desired variance of y}

\item{iter}{The number of iterations to run to estimate the residual variance.}
}
\value{
Value of residual variance given proposed model
}
\description{
find_sig: Find residual variance of your outcome given a vector of coefficients
}
