% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_response_surface_x.R
\name{gen_response_surf_x}
\alias{gen_response_surf_x}
\title{gen_response_surface_x: Generate IVs for Response Surface}
\usage{
gen_response_surf_x(n, cor_mat, x_names = NULL)
}
\arguments{
\item{n}{Number of observations to generate}

\item{cor_mat}{Correlation matrix between IVs}

\item{x_names}{Names of x variables (optional)}
}
\value{
Data frame of IVs with requested column names.
}
\description{
gen_response_surface_x: Generate IVs for Response Surface
}
