% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ly_surf.R
\name{plot_ly_surf}
\alias{plot_ly_surf}
\title{Title}
\usage{
plot_ly_surf(obj = NULL, max.x = NULL, min.x = NULL, max.y = NULL,
  min.y = NULL, inc = NULL, xlab = NULL, ylab = NULL,
  zlab = NULL, showscale = FALSE)
}
\arguments{
\item{obj}{object created by resp_surf function}

\item{max.x}{max possible value of x. Defines limits.}

\item{min.x}{min possible value of x. Defines limits.}

\item{max.y}{max possible value of y. Defines limits.}

\item{min.y}{max possible value of y. Defines limits.}

\item{inc}{Increments of data to be plotted.}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{zlab}{label for z axis}

\item{showscale}{}
}
\description{
Title
}
