% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_surf.R
\name{resp_surf}
\alias{resp_surf}
\title{resp_surf: Robust response surface analysis}
\usage{
resp_surf(dep_var = NULL, fit_var = NULL, control = NULL,
  data = NULL, robust = FALSE, cluster = NULL)
}
\arguments{
\item{dep_var}{Character value containing the name of response surface analysis (character value)}

\item{fit_var}{Character vector containing the names of fit variables for response surface analysis}

\item{control}{Character vector containing the names of the covariates included in the model}

\item{data}{data frame containing all variables  in the model}

\item{robust}{Should robust standard errors be used?}

\item{cluster}{If clustered, what is the ID variable associated with the cluster?}
}
\description{
resp_surf: Robust response surface analysis
}
